Change 'ExpiresIn' type to int64

+++ internal/domain/session/dao/model.go
@@ -6,6 +6,6 @@ type SessionStorage struct {
 	ID        int
 	Token     string
 	CreatedAt time.Time
-	ExpiresIn int
+	ExpiresIn int64
 	UserID    int
 }
